Move DataUpdateCoordinator to coordinator module in nsw_fuel_station
